+++ elwikijsp.root/bundles_elwiki/org.elwiki.data.persist/src/org/elwiki/data/persist/json/converters/AttachmentContentConverter.java
@@ -51,7 +51,7 @@ public class AttachmentContentConverter extends DeserialiseStuff
 
 		AttachmentContent attachmentContent = Elwiki_dataFactory.eINSTANCE.createAttachmentContent();
 
-		short version = getShort(jsonObject, VERSION);
+		int version = getInt(jsonObject, VERSION);
 		Date creationDate = getDate(jsonObject, CREATION_DATE, context);
 		String author = getString(jsonObject, AUTHOR);
 		String changeNote = getString(jsonObject, CHANGE_NOTE);
